Refactor: Traitから Utilに変更
クラスの内容からしてUtilのほうがいい
