+++ .env.example
@@ -99,6 +99,7 @@ HTTPPROXYTUNNEL=false
 PROXYPORT=
 PROXY=
 PROXYUSERPWD=
+CURL_TIMEOUT=
 
 #OPTION_BATCH_SCHEDULE="signatureA|hh:mm,signatureB|hh:mm,･･･" *hh:mm(0~24h)
 #(ex)OPTION_BATCH_SCHEDULE="TestBatchA|02:00,TestBatchB|03:00"
